Project update v2

+++ src/main/java/com/codeinside/attendancesystem/service/impl/GroupServiceImpl.java
@@ -3,18 +3,13 @@ package com.codeinside.attendancesystem.service.impl;
 import com.codeinside.attendancesystem.dto.request.RequestGroupDto;
 import com.codeinside.attendancesystem.dto.response.ResponseGroupDto;
 import com.codeinside.attendancesystem.entity.Group;
-import com.codeinside.attendancesystem.entity.Student;
 import com.codeinside.attendancesystem.exception.GroupNotFoundException;
-import com.codeinside.attendancesystem.exception.OutOfNumberOfStudentsException;
-import com.codeinside.attendancesystem.exception.OutOfRangeAgeException;
-import com.codeinside.attendancesystem.exception.StudentNotFoundException;
 import com.codeinside.attendancesystem.mapper.GroupMapper;
 import com.codeinside.attendancesystem.repository.GroupRepository;
 import com.codeinside.attendancesystem.repository.StudentRepository;
 import com.codeinside.attendancesystem.service.GroupService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -23,13 +18,11 @@ import java.util.Optional;
 public class GroupServiceImpl implements GroupService {
 
     private final GroupRepository groupRepository;
-    private final StudentRepository studentRepository;
     private final GroupMapper groupMapper;
 
     @Autowired
-    public GroupServiceImpl(GroupRepository groupRepository, StudentRepository studentRepository, GroupMapper groupMapper) {
+    public GroupServiceImpl(GroupRepository groupRepository, GroupMapper groupMapper) {
         this.groupRepository = groupRepository;
-        this.studentRepository = studentRepository;
         this.groupMapper = groupMapper;
     }
 
@@ -41,12 +34,11 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public ResponseGroupDto getGroup(Long groupId) {
-        Optional<Group> group = groupRepository.findById(groupId);
-        ResponseGroupDto responseGroupDto = new ResponseGroupDto();
-        if(group.isPresent()) {
-            responseGroupDto = groupMapper.groupToResponseGroupDto(group.get());
+        Optional<Group> groupOptional = groupRepository.findById(groupId);
+        if(groupOptional.isEmpty()) {
+            throw new GroupNotFoundException();
         }
-        return responseGroupDto;
+        return groupMapper.groupToResponseGroupDto(groupOptional.get());
     }
 
     @Override
@@ -55,12 +47,15 @@ public class GroupServiceImpl implements GroupService {
         if(groupList.isEmpty()) {
             throw new GroupNotFoundException();
         }
-        return groupMapper.GroupListToResponseGroupListDto(groupList);
+        return groupMapper.GroupsToResponseGroupDtos(groupList);
     }
 
     @Override
     public void deleteGroup(Long id) {
-        getGroup(id);
+        Optional<Group> groupOptional = groupRepository.findById(id);
+        if(groupOptional.isEmpty()) {
+            throw new GroupNotFoundException();
+        }
         groupRepository.deleteById(id);
     }
